improved test cases
