+++ tests/test_fbr.py
@@ -7,6 +7,10 @@ import warnings
 
 class TestFeaturesBatchRetrieve(BaseTestCase):
 
+    def __init__(self, *args):
+        super().__init__(*args)
+        self.method = 'fat'
+
     def test_compile_to_fbr(self):
         indexref = pt.IndexRef.of(self.here + "/fixtures/index/data.properties")
         # we only want a candidate set of 2 documents
@@ -23,48 +27,48 @@ class TestFeaturesBatchRetrieve(BaseTestCase):
         res4 = compiled(input)
         
 
-    def test_fbr_reranking(self):
+    def test_fbr_reranking2(self):
         if not pt.terrier.check_version("5.4"):
             self.skipTest("Requires Terrier 5.4")
-        # this test examines the use of ScoringMatchingWithFat 
+        # this test examines the use of ScoringMatchingWithFat, using a particular case known to with Terrier 5.3 
         JIR = pt.java.autoclass('org.terrier.querying.IndexRef')
         indexref = JIR.of(self.here + "/fixtures/index/data.properties")
-        # we only want a candidate set of 2 documents
-        firstpass = pt.BatchRetrieve(indexref, wmodel="BM25") % 2
-        pipe = firstpass >> pt.FeaturesBatchRetrieve(indexref, features=["WMODEL:DPH", "WMODEL:PL2"])
-        input = pd.DataFrame([["1", "Stability"]], columns=['qid', 'query'])
-
-        result0 = firstpass(input)
-        result0_map = { row.docno : row.score for row in result0.itertuples() }
-
-        result = pipe.transform(input)
-        self.assertTrue("qid" in result.columns)
-        self.assertTrue("docno" in result.columns)
-        self.assertTrue("score" in result.columns)
-        self.assertTrue("features" in result.columns)
-        self.assertEqual(2, len(result))
-        self.assertEqual(result.iloc[0]["features"].size, 2)
+        # we only want a candidate set of 3 documents
+        firstpass = pt.BatchRetrieve(indexref, wmodel="BM25") % 3
+        pipe1 = firstpass >> pt.FeaturesBatchRetrieve(indexref, features=["WMODEL:PL2", "WMODEL:Tf"], method=self.method)
+        pipe2 = firstpass >> pt.BatchRetrieve(indexref, wmodel="PL2")
+        pipe3 = firstpass >> pt.BatchRetrieve(indexref, wmodel="Tf")
         
-        result1S_map = { row.docno : row.score for row in result.itertuples() }
-        self.assertEqual(result0_map, result1S_map)
+        input = pd.DataFrame([["1", "Stability"]], columns=['qid', 'query'])
+        result0 = firstpass.transform(input)
+        result1 = pipe1.transform(input)
+        result2 = pipe2.transform(input)
+        result3 = pipe3.transform(input)
 
-        pipe_simple = firstpass >> (pt.BatchRetrieve(indexref, wmodel="DPH") ** pt.BatchRetrieve(indexref, wmodel="PL2"))
-        result2 = pipe.transform(input)
-        import numpy as np
-        f1 = np.stack(result["features"].values)
-        f2 = np.stack(result2["features"].values)
-        self.assertTrue( np.array_equal(f1, f2))
+        print(result1)
+        print(result3)
+        #BM25 score
+        result0_map = { row.docno : row.score for row in result0.itertuples() }
+        result1S_map = { row.docno : row.score for row in result1.itertuples() }
+        #Tf score
+        result12F_map = { row.docno : row.features[1] for row in result1.itertuples() }
+        result3_map = { row.docno : row.score for row in result3.itertuples() }
+        #PL2 score
+        result1F_map = { row.docno : row.features[0] for row in result1.itertuples() }
+        result2_map = { row.docno : row.score for row in result2.itertuples() }
 
-        result2S_map = { row.docno : row.score for row in result2.itertuples() }
-        self.assertEqual(result0_map, result2S_map)
+        
 
-        result1F0_map = { row.docno : row.features[0] for row in result.itertuples() }
-        result2F0_map = { row.docno : row.features[0] for row in result2.itertuples() }
-        result1F1_map = { row.docno : row.features[1] for row in result.itertuples() }
-        result2F1_map = { row.docno : row.features[1] for row in result2.itertuples() }
-
-        self.assertEqual(result1F0_map, result2F0_map)
-        self.assertEqual(result1F1_map, result2F1_map)
+        # check features scores
+        # NB: places can go no less than 4, as two documents have similar PL2 scores
+        for rank, row in enumerate(result0.itertuples()):
+            docno = row.docno
+            # check that score is unchanged
+            self.assertAlmostEqual(result1S_map[docno], result0_map[docno], msg="input score mismatch at rank %d for docno %s" % (rank, docno), places=4)
+            #  check that Tf feature score is correct
+            self.assertAlmostEqual(result12F_map[docno], result3_map[docno], msg="Tf feature score mismatch at rank %d for docno %s" % (rank, docno))
+            #  check that feature score is correct
+            self.assertAlmostEqual(result1F_map[docno], result2_map[docno], msg="PL2 feature score mismatch at rank %d for docno %s" % (rank, docno), places=4)
 
     def test_fbr_reranking2(self):
         if not pt.terrier.check_version("5.4"):
@@ -74,7 +78,7 @@ class TestFeaturesBatchRetrieve(BaseTestCase):
         indexref = JIR.of(self.here + "/fixtures/index/data.properties")
         # we only want a candidate set of 3 documents
         firstpass = pt.BatchRetrieve(indexref, wmodel="BM25") % 3
-        pipe1 = firstpass >> pt.FeaturesBatchRetrieve(indexref, features=["WMODEL:PL2"])
+        pipe1 = firstpass >> pt.FeaturesBatchRetrieve(indexref, features=["WMODEL:PL2"], method=self.method)
         pipe2 = firstpass >> pt.BatchRetrieve(indexref, wmodel="PL2")
         
         input = pd.DataFrame([["1", "Stability"]], columns=['qid', 'query'])
@@ -106,7 +110,7 @@ class TestFeaturesBatchRetrieve(BaseTestCase):
     def test_fbr_ltr(self):
         JIR = pt.java.autoclass('org.terrier.querying.IndexRef')
         indexref = JIR.of(self.here + "/fixtures/index/data.properties")
-        retr = pt.FeaturesBatchRetrieve(indexref, ["WMODEL:PL2"], wmodel="DPH")
+        retr = pt.FeaturesBatchRetrieve(indexref, ["WMODEL:PL2"], wmodel="DPH", method=self.method)
         topics = pt.io.read_topics(self.here + "/fixtures/vaswani_npl/query-text.trec").head(3)
         qrels = pt.io.read_qrels(self.here + "/fixtures/vaswani_npl/qrels")
         res = retr.transform(topics)
@@ -119,7 +123,7 @@ class TestFeaturesBatchRetrieve(BaseTestCase):
     def test_fbr(self):
         JIR = pt.java.autoclass('org.terrier.querying.IndexRef')
         indexref = JIR.of(self.here + "/fixtures/index/data.properties")
-        retr = pt.FeaturesBatchRetrieve(indexref, ["WMODEL:PL2"], wmodel="DPH")
+        retr = pt.FeaturesBatchRetrieve(indexref, ["WMODEL:PL2"], wmodel="DPH", method=self.method)
         input = pd.DataFrame([["1", "Stability"]], columns=['qid', 'query'])
         result = retr.transform(input)
         self.assertTrue("qid" in result.columns)
@@ -145,26 +149,38 @@ class TestFeaturesBatchRetrieve(BaseTestCase):
         BM25 = pt.BatchRetrieve(index, wmodel="BM25")
 
         # these rankers we will use to re-rank the BM25 results
-        TF_IDF =  pt.BatchRetrieve(index, wmodel="Dl")
+        Tf =  pt.BatchRetrieve(index, wmodel="Tf")
+        Dl =  pt.BatchRetrieve(index, wmodel="Dl")
         PL2 =  pt.BatchRetrieve(index, wmodel="PL2")
 
-        pipe =  (BM25 %2) >> (TF_IDF ** PL2)
-        fbr = pt.FeaturesBatchRetrieve(indexref, ["WMODEL:Dl", "WMODEL:PL2"], wmodel="BM25") % 2
+        pipe =  (BM25 %2) >> (Tf ** Dl ** PL2)
         resultP = pipe.search("chemical")
-        resultF = fbr.search("chemical")
-        pd.set_option('display.max_columns', None)
+        
+        variants = [
+            pt.FeaturesBatchRetrieve(indexref, ["WMODEL:Tf", "WMODEL:Dl", "WMODEL:PL2"], wmodel="BM25", method=self.method) % 2,
+            BM25 %2 >> pt.FeaturesBatchRetrieve(indexref, ["WMODEL:Tf", "WMODEL:Dl", "WMODEL:PL2"], method=self.method) 
+        ]
 
-        self.assertEqual(resultP.iloc[0].docno, resultF.iloc[0].docno)
-        self.assertEqual(resultP.iloc[0].score, resultF.iloc[0].score)
-        self.assertEqual(resultP.iloc[0].features[0], resultF.iloc[0].features[0])
-        self.assertEqual(resultP.iloc[0].features[1], resultF.iloc[0].features[1])
+        for fbr in variants:
+            resultF = fbr.search("chemical")
+            print(resultP)
+            print(resultF)
+            
+            pd.set_option('display.max_columns', None)
 
-        pipeCompiled = pipe.compile()
-        resultC = pipeCompiled.search("chemical")
-        self.assertEqual(resultP.iloc[0].docno, resultC.iloc[0].docno)
-        self.assertEqual(resultP.iloc[0].score, resultC.iloc[0].score)
-        self.assertEqual(resultP.iloc[0].features[0], resultC.iloc[0].features[0])
-        self.assertEqual(resultP.iloc[0].features[1], resultC.iloc[0].features[1])
+            self.assertEqual(resultP.iloc[0].docno, resultF.iloc[0].docno)
+            self.assertEqual(resultP.iloc[0].score, resultF.iloc[0].score)
+            self.assertEqual(resultP.iloc[0].features[0], resultF.iloc[0].features[0])
+            self.assertEqual(resultP.iloc[0].features[1], resultF.iloc[0].features[1])
+            self.assertEqual(resultP.iloc[0].features[2], resultF.iloc[0].features[2])
+
+            pipeCompiled = pipe.compile()
+            resultC = pipeCompiled.search("chemical")
+            self.assertEqual(resultP.iloc[0].docno, resultC.iloc[0].docno)
+            self.assertEqual(resultP.iloc[0].score, resultC.iloc[0].score)
+            self.assertEqual(resultP.iloc[0].features[0], resultC.iloc[0].features[0])
+            self.assertEqual(resultP.iloc[0].features[1], resultC.iloc[0].features[1])
+            self.assertEqual(resultP.iloc[0].features[2], resultC.iloc[0].features[2])
 
     def test_fbr_empty(self):
         JIR = pt.java.autoclass('org.terrier.querying.IndexRef')
